+++ text-runner/src/commands/dynamic.ts
@@ -1,6 +1,6 @@
 import * as color from "colorette"
 import { extractActivities } from "../activity-list/extract-activities"
-import { instantiateFormatter } from "../configuration/instantiate-formatter"
+import { instantiateFormatter } from "../formatters/instantiate"
 import { getFileNames } from "../filesystem/get-filenames"
 import { findLinkTargets } from "../link-targets/find-link-targets"
 import { parseMarkdownFiles } from "../parsers/markdown/parse-markdown-files"
