+++ text-runner/src/commands/static.ts
@@ -1,6 +1,6 @@
 import * as color from "colorette"
 import { extractImagesAndLinks } from "../activity-list/extract-images-and-links"
-import { instantiateFormatter } from "../configuration/instantiate-formatter"
+import { instantiateFormatter } from "../formatters/instantiate"
 import { getFileNames } from "../filesystem/get-filenames"
 import { findLinkTargets } from "../link-targets/find-link-targets"
 import { parseMarkdownFiles } from "../parsers/markdown/parse-markdown-files"
