+++ text-runner/src/configuration/types/configuration.ts
@@ -1,5 +1,5 @@
 import { Publications } from "../publications/publications"
-import { FormatterName } from "../../formatters/types/formatter-name"
+import { FormatterName } from "../../formatters/formatter-name"
 
 /**
  * All configuration values that are provided to the application
