+++ text-runner/src/formatters/instantiate.ts
@@ -1,11 +1,11 @@
 import { UserError } from "../errors/user-error"
-import { DetailedFormatter } from "../formatters/detailed-formatter"
-import { DotFormatter } from "../formatters/dot-formatter"
-import { ProgressFormatter } from "../formatters/progress-formatter"
-import { SummaryFormatter } from "../formatters/summary-formatter"
-import { Formatter } from "../formatters/types/formatter"
-import { Configuration } from "./types/configuration"
-import { SilentFormatter } from "../formatters/silent-formatter"
+import { DetailedFormatter } from "./types/detailed-formatter"
+import { DotFormatter } from "./types/dot-formatter"
+import { ProgressFormatter } from "./progress-formatter"
+import { SummaryFormatter } from "./types/summary-formatter"
+import { Formatter } from "./formatter"
+import { Configuration } from "../configuration/types/configuration"
+import { SilentFormatter } from "./types/silent-formatter"
 
 export function instantiateFormatter(
   name: string | undefined,
