+++ text-runner/src/formatters/progress-formatter.ts
@@ -4,8 +4,8 @@ import * as path from "path"
 import { Activity } from "../activity-list/types/activity"
 import { Configuration } from "../configuration/types/configuration"
 import { printCodeFrame } from "../helpers/print-code-frame"
-import { Formatter } from "./types/formatter"
-import { printSummary } from "./shared/print-summary"
+import { Formatter } from "./formatter"
+import { printSummary } from "./print-summary"
 import { StatsCounter } from "../runners/helpers/stats-counter"
 
 export class ProgressFormatter implements Formatter {
