+++ text-runner/src/formatters/types/detailed-formatter.ts
@@ -1,11 +1,11 @@
 import * as color from "colorette"
 import * as path from "path"
-import { Activity } from "../activity-list/types/activity"
-import { Configuration } from "../configuration/types/configuration"
-import { printCodeFrame } from "../helpers/print-code-frame"
-import { Formatter } from "./types/formatter"
-import { printSummary } from "./shared/print-summary"
-import { StatsCounter } from "../runners/helpers/stats-counter"
+import { Activity } from "../../activity-list/types/activity"
+import { Configuration } from "../../configuration/types/configuration"
+import { printCodeFrame } from "../../helpers/print-code-frame"
+import { Formatter } from "../formatter"
+import { printSummary } from "../print-summary"
+import { StatsCounter } from "../../runners/helpers/stats-counter"
 
 /** A formatter that prints output and step names */
 export class DetailedFormatter implements Formatter {
