+++ text-runner/src/formatters/types/dot-formatter.ts
@@ -1,11 +1,11 @@
 import * as color from "colorette"
 import * as path from "path"
-import { Activity } from "../activity-list/types/activity"
-import { Configuration } from "../configuration/types/configuration"
-import { printCodeFrame } from "../helpers/print-code-frame"
-import { Formatter } from "./types/formatter"
-import { StatsCounter } from "../runners/helpers/stats-counter"
-import { printSummary } from "./shared/print-summary"
+import { Activity } from "../../activity-list/types/activity"
+import { Configuration } from "../../configuration/types/configuration"
+import { printCodeFrame } from "../../helpers/print-code-frame"
+import { Formatter } from "../formatter"
+import { StatsCounter } from "../../runners/helpers/stats-counter"
+import { printSummary } from "../print-summary"
 
 /** A minimalistic formatter, prints dots for each check */
 export class DotFormatter implements Formatter {
