+++ text-runner/src/formatters/types/summary-formatter.ts
@@ -1,11 +1,11 @@
 import * as color from "colorette"
 import * as path from "path"
-import { Activity } from "../activity-list/types/activity"
-import { Configuration } from "../configuration/types/configuration"
-import { printCodeFrame } from "../helpers/print-code-frame"
-import { Formatter } from "./types/formatter"
-import { printSummary } from "./shared/print-summary"
-import { StatsCounter } from "../runners/helpers/stats-counter"
+import { Activity } from "../../activity-list/types/activity"
+import { Configuration } from "../../configuration/types/configuration"
+import { printCodeFrame } from "../../helpers/print-code-frame"
+import { Formatter } from "../formatter"
+import { printSummary } from "../print-summary"
+import { StatsCounter } from "../../runners/helpers/stats-counter"
 
 /** An extremely minimalistic formatter, prints only a summary at the end */
 export class SummaryFormatter implements Formatter {
