+++ text-runner/src/runners/execute-sequential.ts
@@ -1,6 +1,6 @@
 import { ActivityList } from "../activity-list/types/activity-list"
 import { Configuration } from "../configuration/types/configuration"
-import { Formatter } from "../formatters/types/formatter"
+import { Formatter } from "../formatters/formatter"
 import { LinkTargetList } from "../link-targets/link-target-list"
 import { StatsCounter } from "./helpers/stats-counter"
 import { runActivity } from "./run-activity"
