+++ text-runner/src/runners/run-activity.ts
@@ -6,7 +6,7 @@ import { ActionArgs } from "../actions/types/action-args"
 import { ActionResult } from "../actions/types/action-result"
 import { Activity } from "../activity-list/types/activity"
 import { Configuration } from "../configuration/types/configuration"
-import { Formatter } from "../formatters/types/formatter"
+import { Formatter } from "../formatters/formatter"
 import { LinkTargetList } from "../link-targets/link-target-list"
 import { NameRefiner } from "./helpers/name-refiner"
 import { OutputCollector } from "./helpers/output-collector"
